chore: split migration modes into separate folders
